+++ internal/retryprovenance/keys.go
@@ -1,10 +1,11 @@
 package retryprovenance
 
 const (
-	RepoDirKey      = "_SPARKWING_RETRY_REPO_DIR"
-	RepoIdentityKey = "_SPARKWING_RETRY_REPO_URL"
-	RevisionKey     = "_SPARKWING_RETRY_REVISION"
-	PlanHashKey     = "_SPARKWING_RETRY_PLAN_HASH"
+	PipelineRevisionKey = "_SPARKWING_SUBMIT_PIPELINE_REV"
+	RepoDirKey          = "_SPARKWING_RETRY_REPO_DIR"
+	RepoIdentityKey     = "_SPARKWING_RETRY_REPO_URL"
+	RevisionKey         = "_SPARKWING_RETRY_REVISION"
+	PlanHashKey         = "_SPARKWING_RETRY_PLAN_HASH"
 
 	RecordedRevisionSnapshotPolicy = "recorded_revision_snapshot"
 )